requirements.txt
pycryptodome-test-vectors==1.0.4
pycryptodomex==3.12.0
